+++ src/components/Dialog/index.ts
@@ -3,4 +3,5 @@ export * from "./FormDialog";
 export * from "./ConfirmationDialog";
 export * from "./DialogActions";
 export * from "./ImportDialog";
+export * from "./ExportDialog";
 export * from "./types";