+++ src/hooks/dialogs/index.ts
@@ -5,4 +5,5 @@ export * from "./usePostDialog";
 export * from "./usePutDialog";
 export * from "./useRestoreDialog";
 export * from "./useImportDialog";
+export * from "./useExportDialog";
 export type * from "./types";